requirements.txt
api==0.0.7
asgiref==3.6.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.0.3
django-allauth==0.57.0
django-cors-headers==4.0.0
django-environ==0.11.2
django-webpack-loader==1.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environ==1.0
idna==3.4
mysqlclient==2.2.0
nose==1.3.7
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.2
